Dll layering - Host.UnitTests and TestCommon should not depend on Storage/ServiceBus. Move storage tests from Unit into Functional.
No product change.
