+++ test/Microsoft.Azure.WebJobs.Host.FunctionalTests/DataBindingProviderTests.cs
