+++ test/Microsoft.Azure.WebJobs.Host.FunctionalTests/Indexers/FunctionIndexerIntegrationErrorTests.cs
