+++ test/Microsoft.Azure.WebJobs.Host.FunctionalTests/Loggers/FunctionStartedMessageExtensionsTests.cs
