+++ test/Microsoft.Azure.WebJobs.Host.FunctionalTests/Loggers/UpdateOutputLogCommandExtensions.cs
