+++ test/Microsoft.Azure.WebJobs.Host.FunctionalTests/Loggers/UpdateOutputLogCommandTests.cs
