+++ test/Microsoft.Azure.WebJobs.Host.FunctionalTests/Singleton/SingletonConfigurationTests.cs
