+++ test/Microsoft.Azure.WebJobs.Host.FunctionalTests/Singleton/SingletonEnd2EndTests.cs
