+++ test/Microsoft.Azure.WebJobs.Host.FunctionalTests/Singleton/SingletonListenerTests.cs
