+++ test/Microsoft.Azure.WebJobs.Host.FunctionalTests/Singleton/SingletonLockTests.cs
