+++ test/Microsoft.Azure.WebJobs.Host.FunctionalTests/Singleton/SingletonValueProviderTests.cs
