+++ test/Microsoft.Azure.WebJobs.Host.FunctionalTests/TypeUtilityTests.cs
