+++ test/Microsoft.Azure.WebJobs.Host.FunctionalTests/appsettings.json
@@ -1,0 +1,7 @@
+﻿{
+  "DisableSetting0": "0",
+  "DisableSetting1": "1",
+  "DisableSettingTrue": "true",
+  "DisableSettingFalse": "false",
+  "AzureWebJobs.DisabledByAppSetting_NoAttribute.Disabled": "true"  
+}
