+++ test/Microsoft.Azure.WebJobs.Host.UnitTests/JobHostMetadataProviderTests.cs
@@ -74,120 +74,6 @@ namespace Microsoft.Azure.WebJobs.Host.UnitTests
             return (T)attribute;
         }
 
-        [Fact]
-        public void AttrBuilder()
-        {
-            IHost host = new HostBuilder()
-                .ConfigureDefaultTestHost()
-                .Build();
-            
-            var metadataProvider = host.CreateMetadataProvider();
-
-            // Blob 
-            var blobAttr = GetAttr<BlobAttribute>(metadataProvider, new { path = "x" });
-            Assert.Equal("x", blobAttr.BlobPath);
-            Assert.Equal(null, blobAttr.Access);
-
-            blobAttr = GetAttr<BlobAttribute>(metadataProvider, new { path = "x", direction = "in" });
-            Assert.Equal("x", blobAttr.BlobPath);
-            Assert.Equal(FileAccess.Read, blobAttr.Access);
-
-            blobAttr = GetAttr<BlobAttribute>(metadataProvider, new { Path = "x", Direction = "out" });
-            Assert.Equal("x", blobAttr.BlobPath);
-            Assert.Equal(FileAccess.Write, blobAttr.Access);
-
-            blobAttr = GetAttr<BlobAttribute>(metadataProvider, new { path = "x", direction = "inout" });
-            Assert.Equal("x", blobAttr.BlobPath);
-            Assert.Equal(FileAccess.ReadWrite, blobAttr.Access);
-
-            blobAttr = GetAttr<BlobAttribute>(metadataProvider,
-            new
-            {
-                path = "x",
-                direction = "in",
-                connection = "cx1"
-            });
-            Assert.Equal("x", blobAttr.BlobPath);
-            Assert.Equal(FileAccess.Read, blobAttr.Access);
-            Assert.Equal("cx1", blobAttr.Connection);
-
-            blobAttr = GetAttr<BlobAttribute>(metadataProvider,
-              new
-              {
-                  path = "x",
-                  direction = "in",
-                  connection = "" // empty, not null 
-              });
-            Assert.Equal("x", blobAttr.BlobPath);
-            Assert.Equal(FileAccess.Read, blobAttr.Access);
-            Assert.Equal("", blobAttr.Connection); // empty is passed straight through. 
-
-            var blobTriggerAttr = GetAttr<BlobTriggerAttribute>(metadataProvider, new { path = "x" });
-            Assert.Equal("x", blobTriggerAttr.BlobPath);
-
-            // Queue 
-            var queueAttr = GetAttr<QueueAttribute>(metadataProvider, new { QueueName = "q1" });
-            Assert.Equal("q1", queueAttr.QueueName);
-
-            var queueTriggerAttr = GetAttr<QueueTriggerAttribute>(metadataProvider, new { QueueName = "q1" });
-            Assert.Equal("q1", queueTriggerAttr.QueueName);
-
-            // Table
-            var tableAttr = GetAttr<TableAttribute>(metadataProvider, new { TableName = "t1" });
-            Assert.Equal("t1", tableAttr.TableName);
-
-            tableAttr = GetAttr<TableAttribute>(metadataProvider, new { TableName = "t1", partitionKey = "pk", Filter = "f1" });
-            Assert.Equal("t1", tableAttr.TableName);
-            Assert.Equal("pk", tableAttr.PartitionKey);
-            Assert.Equal(null, tableAttr.RowKey);
-            Assert.Equal("f1", tableAttr.Filter);
-        }
-
-        [Fact]
-        public void DefaultTypeForTable()
-        {
-            var host = new HostBuilder()
-                .ConfigureDefaultTestHost()
-                .AddAzureStorage()
-                .Build();
-
-            var metadataProvider = host.CreateMetadataProvider();
-
-            var t1 = metadataProvider.GetDefaultType(new TableAttribute("table1"), FileAccess.Read, null);
-            Assert.Equal(typeof(JArray), t1);
-
-            var t2 = metadataProvider.GetDefaultType(new TableAttribute("table1", "pk", "rk"), FileAccess.Read, null);
-            Assert.Equal(typeof(JObject), t2);
-
-            var t3 = metadataProvider.GetDefaultType(new TableAttribute("table1"), FileAccess.Write, null);
-            Assert.Equal(typeof(IAsyncCollector<JObject>), t3);
-        }
-
-
-        [Fact]
-        public void DefaultTypeForQueue()
-        {
-            var host = new HostBuilder()
-                .ConfigureDefaultTestHost()
-                .Build();
-
-            var metadataProvider = host.CreateMetadataProvider();
-
-            var t1 = metadataProvider.GetDefaultType(new QueueTriggerAttribute("q"), FileAccess.Read, typeof(byte[]));
-            Assert.Equal(typeof(byte[]), t1);
-
-            var t2 = metadataProvider.GetDefaultType(new QueueTriggerAttribute("q"), FileAccess.Read, null);
-            Assert.Equal(typeof(string), t2);
-
-            // Very important that this is byte[]. 
-            // Script doesn't require Function.json for JScript to specify datatype. 
-            // JScript can convert Jobject, string to byte[].
-            // But can't convert byte[] to JObject. 
-            // so byte[] is the safest default. 
-            var t3 = metadataProvider.GetDefaultType(new QueueAttribute("q"), FileAccess.Write, null);
-            Assert.Equal(typeof(IAsyncCollector<byte[]>), t3);
-        }
-
         // This is a setup used by CosmoDb. 
         [Fact]
         public void DefaultTypeForOpenTypeCollector()
